+++ src/lib/Components/Inbox/Conversations/__tests__/Composer-tests.tsx
@@ -19,9 +19,11 @@ describe("regarding the send button", () => {
 
   it("disables it even if it contains text if the disabled prop is true", () => {
     const overrideText = "History repeats itself, first as tragedy, second as farce."
-    const tree = renderer.create(<Composer value={overrideText} disabled={true} />) as any
+    // We're using 'dive' here to only fetch the component we want to test
+    // This is because the component is wrapped by react-tracking, which changes the tree structure
+    const tree = shallow(<Composer value={overrideText} disabled={true} />).dive()
 
-    const instance = tree.getInstance()
+    const instance = tree.instance()
     instance.componentDidUpdate()
 
     expect(tree).toMatchSnapshot()
@@ -29,7 +31,9 @@ describe("regarding the send button", () => {
 
   it("calls onSubmit with the text when send button is pressed", () => {
     const onSubmit = jest.fn()
-    const wrapper = shallow(<Composer onSubmit={onSubmit} />)
+    // We're using 'dive' here to only fetch the component we want to test
+    // This is because the component is wrapped by react-tracking, which changes the tree structure
+    const wrapper = shallow(<Composer onSubmit={onSubmit} />).dive()
     const text = "Don't trust everything you see, even salt looks like sugar"
     wrapper.setState({ text })
     wrapper.find(TouchableWithoutFeedback).simulate("press")